refactor: Split project to modules, seperate by 'internal' keyword, add HTLBProxyServer class for testing
