+++ cache/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    id("convention")
+}
+
+dependencies {
+    implementation(libs.kotlinXcoroutines)
+    implementation(libs.commonCollections)
+    implementation(project(":utils"))
+}
