+++ server/src/main/kotlin/proxy/CacheInfo.kt
