--- src/main/kotlin/io/github/darefox/hltbproxy/Json.kt
@@ -1,7 +1,0 @@
-package io.github.darefox.hltbproxy
-
-import kotlinx.serialization.json.Json
-
-val jsonKotlinx = Json {
-    ignoreUnknownKeys = true
-}
