--- src/main/kotlin/io/github/darefox/hltbproxy/Main.kt
@@ -1,11 +1,0 @@
-package io.github.darefox.hltbproxy
-
-import io.github.darefox.hltbproxy.proxy.serverRoutes
-import org.http4k.server.Jetty
-import org.http4k.server.asServer
-
-
-fun main(args: Array<String>) {
-    val port = System.getenv("PORT")?.toIntOrNull() ?: 8080
-    serverRoutes.asServer(Jetty(port)).start()
-}
