--- src/main/kotlin/io/github/darefox/hltbproxy/http4k/client.kt
@@ -1,5 +1,0 @@
-package io.github.darefox.hltbproxy.http4k
-
-import org.http4k.client.ApacheClient
-
-val client = ApacheClient()
