+++ utils/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    id("convention")
+}
+
+dependencies {
+    implementation(libs.kotlinXcoroutines)
+    implementation(libs.commonCollections)
+}
